updated Model classes, Implemented Services to calculate Death Premium and get occupation related information

+++ TAL-PremiumCalculatorAPI/Models/OccupationRating.cs
@@ -3,6 +3,6 @@
     public class OccupationRating
     {
         public string Rating { get; set; }
-        public decimal Factor { get; set; }
+        public double Factor { get; set; }
     }
 }
